Split Etherscan Responses into own mock file
