+++ packages/transaction-controller/src/TransactionControllerIntegration.test.ts
@@ -10,7 +10,7 @@ import nock from 'nock';
 import { mockNetwork } from '../../../tests/mock-network';
 import { TransactionController } from './TransactionController';
 import { getEtherscanApiHost } from './utils/etherscan';
-import { ETHERSCAN_TRANSACTION_RESPONSE_MOCK } from './helpers/EtherscanRemoteTransactionSource.test';
+import { ETHERSCAN_TRANSACTION_RESPONSE_MOCK } from './helpers/EtherscanMocks';
 
 const ACCOUNT_MOCK = '0x6bf137f335ea1b8f193b8f6ea92561a60d23a207';
 const ACCOUNT_2_MOCK = '0x08f137f335ea1b8f193b8f6ea92561a60d23a211';
